Skip to content

Docker Files and Images

With each release we provide a set of Dockerfiles and pre-built Docker images. The Dockerfiles can be used to build, and modify, your own image, while the images can be used directly to use ILLIXR. These images derive from NVIDIA cuda based images, using the latest Ubuntu LTS release.

Prerequisites

The following must be installed for these images to be used:

Docker, you must have an installed and running Docker engine

NVIDIA Container Toolkit, provides access to the local NVIDIA GPU from the container.1

Alternately the pre-built images can be downloaded from . There are several versions available depending on your needs:

Versions

We provide several versions of ILLIXR, depending on your needs:

Name 	Description

Full 	Build all plugins and dependencies.

No GPL 	Build all plugins and dependencies which are not covered by a GPL license.

No LGPL 	Build all plugins and dependencies which are not covered by a LGPL or GPL license.

No ZED 	Build all plugins and dependencies except the ZEDS related ones.

No GPL or ZED 	Build all plugins and dependencies which are not covered by a GPL license, also do not build ZED related code.

No LGPL or ZED 	Build all plugins and dependencies which are not covered by a LGPL or GPL license, also do not build ZED related code.

The images can be constructed with the provided script docksr.sh which takes the following options:

  -a : make all possible images (exclusive with all other flags)
  -f : docker image will have everything (exclusive with all other flags)
  -g : docker image will have no GPL licensed code (exclusive with -a and -f)
  -h : print this help and exit
  -l : docker image will have no LGPL or GPL licensed code, implies -g (exclusive with -a and -f)
  -z : docker image will not have the ZED SDK (exclusive with -a and -f)

In all versions:

the ILLIXR code are git cloned into /home/illixr/ILLIXR

the main user is named illixr with no password

the illixr user has sudo privileges

any binaries (executables) compiled and installed by the build process are installed in /home/illixr/bin

any libraries compiled and installed by the build process are installed in /home/illixr/lib

/home/illixr/bin is already in the user's $PATH

/home/illixr/lib is already in the user's $LD_LIBRARY_PATH

you can start ILLIXR as normal main.dbg.exe ....

Note

The use of the --privileged flag to docker run is recommended for our containers as it gives the container access to all devices on the host machine, allowing for easy use of usb cameras and other devices.

Full Build

Build:

docker.sh -f

Docker image: illixr/illixr_full:latest

Download:

docker pull ghcr.io/illixr/illixr_full:latest

Run:

docker run --gpus all -it --privileged illixr_full:latest</pre>

Docker Compose

export ILLIXR_DOCKER_IMAGE=full
docker compose -f docker-compose.yaml up

No GPL

Build:

docker.sh -g

Docker image: illixr/illixr_no_gpl:latest

Download:

docker pull ghcr.io/illixr/illixr_no_gpl:latest

Run:

docker run --gpus all -it --privileged illixr_no_gpl:latest

Docker Compose

export ILLIXR_DOCKER_IMAGE=no_gpl
docker compose -f docker-compose.yaml up

No LGPL

Build:

docker.sh -l

Docker image: illixr/illixr_no_lgpl:latest

Download:

docker pull ghcr.io/illixr/illixr_no_lgpl:latest

Run:

docker run --gpus all -it --privileged illixr_no_lgpl:latest

Docker Compose

export ILLIXR_DOCKER_IMAGE=no_lgpl
docker compose -f docker-compose.yaml up

No Zed

Build:

docker.sh -z

Docker image: illixr/illixr_no_zed:latest

Download:

docker pull ghcr.io/illixr/illixr_no_zed:latest

Run:

docker run --gpus all -it --privileged illixr_no_zed:latest

Docker Compose

export ILLIXR_DOCKER_IMAGE=no_zed
docker compose -f docker-compose.yaml up

No GPL or ZED

Build:

docker.sh -gz

Docker image: illixr/illixr_no_gpl_zed:latest

Download:

docker pull ghcr.io/illixr/illixr_no_gpl_zed:latest

Run:

docker run --gpus all -it --privileged illixr_no_gpl_zed:latest

Docker Compose

export ILLIXR_DOCKER_IMAGE=no_gpl_zed
docker compose -f docker-compose.yaml up

No LGPL or ZED

Build:

docker.sh -lz

Docker image: illixr/illixr_no_lgpl_zed:latest

Download:

docker pull ghcr.io/illixr/illixr_no_lgpl_zed:latest

Run:

docker run --gpus all -it --privileged illixr_no_lgpl_zed:latest

Docker Compose

export ILLIXR_DOCKER_IMAGE=no_lgpl_zed
docker compose -f docker-compose.yaml up

A common issue with running the NVIDIA images produces the error Failed to initialize NVML: Unknown Error. A solution to this can be found here ↩