Skip to content

Class Hierarchy

This inheritance list is sorted roughly, but not completely, alphabetically:

class HMD

class ILLIXR::cancellable_sleep

class ILLIXR::phonebook::service A 'service' that can be registered in the phonebook.

class ILLIXR::data_format::pose_prediction

class ILLIXR::fauxpose_impl Create a "pose_prediction" type service.

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::pose_lookup_impl

class ILLIXR::pose_prediction_impl

class ILLIXR::gen_guid This class generates unique IDs.

class ILLIXR::network::network_backend

class ILLIXR::tcp_network_backend

class ILLIXR::record_logger The ILLIXR logging service for structured records.

class ILLIXR::sqlite_record_logger

class ILLIXR::stdout_record_logger

class ILLIXR::relative_clock Relative clock for all of ILLIXR .

class ILLIXR::stoplight Start/stop synchronization for the whole application.

class ILLIXR::switchboard A manager for typesafe, threadsafe, named event-streams (called topics).

class ILLIXR::vulkan::display_provider A display sink is a service that can display the rendered images to the screen.

class ILLIXR::vulkan::render_pass

class ILLIXR::vulkan::app

class ILLIXR::offload_rendering_client

class ILLIXR::vkdemo

class ILLIXR::vulkan::timewarp

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::openwarp_vk

class ILLIXR::timewarp_vk

class ILLIXR::xlib_gl_extended_window

class ILLIXR::plugin A dynamically-loadable plugin for Spindle.

class ILLIXR::depthai

class ILLIXR::fauxpose

class ILLIXR::ground_truth_slam

class ILLIXR::gtsam_integrator

class ILLIXR::offload_data

class ILLIXR::offload_rendering_server_loader Plugin loader for the offload rendering server.

class ILLIXR::passthrough_integrator

class ILLIXR::realsense

class ILLIXR::record_imu_cam

class ILLIXR::record_rgb_depth

class ILLIXR::rk4_integrator

class ILLIXR::server_writer

class ILLIXR::tcp_network_backend

class ILLIXR::threadloop A reusable threadloop for plugins.

class ILLIXR::data_injection

class ILLIXR::debugview

class ILLIXR::gldemo

class ILLIXR::lighthouse

class ILLIXR::native_renderer

class ILLIXR::offline_cam

class ILLIXR::offline_imu

class ILLIXR::offload_reader

class ILLIXR::offload_rendering_client

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::offload_writer

class ILLIXR::openni_plugin

class ILLIXR::openwarp_vk_plugin

class ILLIXR::server_reader

class ILLIXR::timewarp_gl

class ILLIXR::timewarp_vk_plugin

class ILLIXR::webcam

class ILLIXR::zed_camera_thread

class ILLIXR::zed_imu_thread

class ILLIXR::viewer

class ILLIXR::vkdemo_plugin

class offload_rendering_client_loader Plugin loader for the offload rendering client.

class pose_lookup_plugin

class pose_prediction_plugin

class ILLIXR::data_use_indicator A helper class that lets one dynamically determine if some data gets used.

class ILLIXR::display::display_backend

class ILLIXR::display::glfw_extended

class ILLIXR::display::headless

class ILLIXR::display::x11_direct

class ILLIXR::dynamic_lib

class ILLIXR::event A boolean condition-variable.

class ILLIXR::lazy_load_image

class ILLIXR::managed_thread An object that manages a std::thread; it joins and exits when the object gets destructed.

class ILLIXR::network::TCPSocket

class ILLIXR::phonebook A service locator forILLIXR .

class ILLIXR::record This class represents a tuple of fields which get logged by record_logger .

class ILLIXR::record_coalescer Coalesces logs of the same type to be written back as a single-transaction.

class ILLIXR::record_header Schema of each record.

class ILLIXR::runtime

class runtime_impl

class ILLIXR::sqlite_thread

class ILLIXR::switchboard::buffered_reader

class ILLIXR::switchboard::event Virtual class for event types.

struct ILLIXR::data_format::cam_base_type

struct ILLIXR::data_format::binocular_cam_type

struct ILLIXR::data_format::cam_type_zed

struct ILLIXR::data_format::depth_type

struct ILLIXR::data_format::ht::ht_frame

struct ILLIXR::data_format::monocular_cam_type

struct ILLIXR::data_format::rgb_depth_type

struct ILLIXR::data_format::camera_data Data structure to hold information about the full camera system. This information is mostly constant.

struct ILLIXR::data_format::compressed_frame

struct ILLIXR::data_format::connection_signal

struct ILLIXR::data_format::fast_pose_type

struct ILLIXR::data_format::hologram_input

struct ILLIXR::data_format::image_handle

struct ILLIXR::data_format::imu_integrator_input

struct ILLIXR::data_format::imu_raw_type

struct ILLIXR::data_format::imu_type

struct ILLIXR::data_format::pose_type

struct ILLIXR::data_format::rendered_frame

struct ILLIXR::data_format::signal_to_quad

struct ILLIXR::data_format::texture_pose

class ILLIXR::switchboard::event_wrapper Helper class for making event types.

class ILLIXR::switchboard::writer A handle which can publish events to a topic.

class ILLIXR::switchboard::network_writer

class ILLIXR::switchboard::network_writer

class ILLIXR::switchboard::network_writer

class ILLIXR::switchboard::reader A handle which can read the latest event on a topic.

class ILLIXR::time_point

class ILLIXR::video_decoder

class ILLIXR::video_encoder

class ILLIXR::zed_capture::capture

class ILLIXR::zed_capture::files

class ILLIXR::gtsam_integrator::pim_object Wrapper object protecting the lifetime of IMU integration inputs and biases.

class ILLIXR::switchboard::coordinate_system Base coordinate system.

class ILLIXR::switchboard::topic Represents a topic.

class ILLIXR::switchboard::topic_buffer

class ILLIXR::switchboard::topic_subscription Represents a single topic_subscription (callback and queue)

class csv_iterator

class csv_row

class cxxopts::KeyValue

class cxxopts::OptionAdder

class cxxopts::OptionDetails

class cxxopts::OptionParser

class cxxopts::OptionValue

class cxxopts::Options

class cxxopts::ParseResult

class cxxopts::ParseResult::Iterator

class offload_rendering_client Main client implementation for offload rendering.

class print_timer Like timer, but prints the output.

class print_timer2

class print_timer::print_in_destructor

class should_profile_class

class timer a timer that times until the end of the code block ([RAII]).

struct HMD::hmd_info_t

struct HMD::mesh_coord2d_t

struct HMD::mesh_coord3d_t

struct HMD::uv_coord_t

struct ILLIXR::Dependency

struct ILLIXR::DistortionCorrectionVertex

struct ILLIXR::DistortionMatrix

struct ILLIXR::OpenWarpVertex

struct ILLIXR::WarpMatrices

struct ILLIXR::data_format::ccd_data A data structure to hold relevant camera information. This information is constant (per camera)

struct ILLIXR::data_format::hmd_physical_info

struct ILLIXR::data_format::ht::ht_detection

struct ILLIXR::data_format::ht::position

struct ILLIXR::data_format::ht::raw_ht_data

struct ILLIXR::data_format::imu_params

struct ILLIXR::data_format::points_with_units

struct ILLIXR::data_format::ht::hand_points

struct ILLIXR::data_format::ht::velocity

struct ILLIXR::data_format::pose_data

struct ILLIXR::data_format::pose_type

struct ILLIXR::data_format::raw_point

struct ILLIXR::data_format::raw_pose

struct ILLIXR::data_format::rect

struct ILLIXR::data_format::vk_image_handle

struct ILLIXR::display_params Display parameters.

struct ILLIXR::index_params

struct ILLIXR::model

struct ILLIXR::network::topic_config

struct ILLIXR::raw_imu_type

struct ILLIXR::rendering_params Rendering parameters.

struct ILLIXR::sensor_types

struct ILLIXR::server_params

struct ILLIXR::state_plus

struct ILLIXR::texture

struct ILLIXR::vertex

struct ILLIXR::vulkan::buffer_pool

struct ILLIXR::vulkan::ffmpeg_utils::ffmpeg_vk_frame

struct ILLIXR::vulkan::queue

struct ILLIXR::vulkan::queue_families

struct ILLIXR::vulkan::swapchain_details

struct ILLIXR::vulkan::vk_image

struct ILLIXR::realsense::accel_type

struct YAML::convert< ILLIXR::Dependency >

struct cxxopts::HelpGroupDetails

struct cxxopts::HelpOptionDetails

struct cxxopts::Option

struct cxxopts::values::detail::SignedCheck

struct cxxopts::values::detail::SignedCheck< T, false >

struct cxxopts::values::detail::SignedCheck< T, true >

struct cxxopts::values::parser_tool::ArguDesc

struct cxxopts::values::parser_tool::IntegerDesc

struct cxxopts::values::type_is_container

struct cxxopts::values::type_is_container< std::vector< T > >

struct model_push_constant

struct std::hash< vertex >

struct uniform_buffer_object

struct vertex

class Eigen::Quaternion< Scalar_, Eigen::AutoAlign >

class ILLIXR::data_format::proper_quaternion A more complete implementation of the Eigen::Quaternion.

class ILLIXR::display::display_backend

class ILLIXR::display::glfw_extended

class ILLIXR::display::headless

class ILLIXR::display::x11_direct

class ILLIXR::data_format::pose_prediction

class ILLIXR::fauxpose_impl Create a "pose_prediction" type service.

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::pose_lookup_impl

class ILLIXR::pose_prediction_impl

class ILLIXR::vulkan::app

class ILLIXR::offload_rendering_client

class ILLIXR::vkdemo

class ILLIXR::vulkan::timewarp

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::openwarp_vk

class ILLIXR::timewarp_vk

class std::enable_shared_from_this< plugin >

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class vulkan::vk_extension_request

class ILLIXR::offload_rendering_server_loader Plugin loader for the offload rendering server.

class offload_rendering_client_loader Plugin loader for the offload rendering client.

class ILLIXR::network::network_backend

class ILLIXR::tcp_network_backend

class ILLIXR::vulkan::render_pass

class ILLIXR::vulkan::app

class ILLIXR::offload_rendering_client

class ILLIXR::vkdemo

class ILLIXR::vulkan::timewarp

class ILLIXR::offload_rendering_server Main server implementation for offload rendering.

class ILLIXR::openwarp_vk

class ILLIXR::timewarp_vk

class sl::Camera

class ILLIXR::zed_camera

class std::enable_shared_from_this< Value >

class cxxopts::Value

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::Value

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class std::exception

class cxxopts::exceptions::exception

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::exception

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::exception

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::option_has_no_value

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::parsing

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::gratuitous_argument_for_option

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::incorrect_argument_type

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::invalid_option_syntax

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::missing_argument

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::no_such_option

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::option_requires_argument

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::requested_option_not_present

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::specification

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::invalid_option_format

class cxxopts::exceptions::option_already_exists

class cxxopts::exceptions::option_already_exists

class cxxopts::Value

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::abstract_value< T >

class cxxopts::values::standard_value

class cxxopts::values::standard_value

class cxxopts::values::abstract_value< bool >

class cxxopts::values::standard_value< bool >

class cxxopts::values::standard_value< bool >

class vulkan::display_provider

class display_vk

class ILLIXR::runtime

class runtime_impl

class ILLIXR::data_format::points_with_units

struct ILLIXR::data_format::ht::hand_points

struct ILLIXR::data_format::ht::velocity

struct ILLIXR::data_format::ht::hand_points

struct ILLIXR::data_format::ht::velocity

class Eigen::Vector3f

struct ILLIXR::data_format::point

struct ILLIXR::data_format::point_with_validity

struct ILLIXR::data_format::point_with_units

class ILLIXR::data_format::point_with_validity

struct ILLIXR::data_format::point_with_units

class ILLIXR::data_format::point

struct ILLIXR::data_format::point_with_validity

struct ILLIXR::data_format::point_with_units