Skip to content

Class Hierarchy

This inheritance list is sorted roughly, but not completely, alphabetically:

class ILLIXR::data_use_indicator A helper class that lets one dynamically determine if some data gets used.

class ILLIXR::dynamic_lib

class ILLIXR::phonebook::service A 'service' that can be registered in the phonebook.

class ILLIXR::gen_guid This class generates unique IDs.

class ILLIXR::record_logger The ILLIXR logging service for structured records.

class ILLIXR::switchboard A manager for typesafe, threadsafe, named event-streams (called topics).

class ILLIXR::xlib_gl_extended_window

class pose_prediction

class ILLIXR::phonebook A service locator for ILLIXR.

class ILLIXR::plugin A dynamically-loadable plugin for Spindle.

class ILLIXR::threadloop A reusable threadloop for plugins.

class ILLIXR::reader_latest A handle which can read the latest event on a topic.

class ILLIXR::record This class represents a tuple of fields which get logged by record_logger .

class ILLIXR::record_coalescer Coalesces logs of the same type to be written back as a single-transaction.

class ILLIXR::record_header Schema of each record.

class ILLIXR::runtime

class ILLIXR::start_end_logger a utility for logging to start and end times to a file

class ILLIXR::writer A handle which can publish events to a topic.

class print_timer Like timer, but prints the output.

class print_timer2

class print_timer::print_in_destructor

class should_profile_class

class timer a timer that times until the end of the code block ([RAII]).

struct ILLIXR::accel

struct ILLIXR::camera_frame

struct ILLIXR::fast_pose_type

struct ILLIXR::hmd_physical_info

struct ILLIXR::hologram_input

struct ILLIXR::hologram_output

struct ILLIXR::imu_cam_type

struct ILLIXR::imu_integrator_input

struct ILLIXR::imu_integrator_seq

struct ILLIXR::imu_params

struct ILLIXR::imu_raw_type

struct ILLIXR::pose_type

struct ILLIXR::rendered_frame

struct ILLIXR::rgb_depth_type

struct linalg::converter

struct linalg::converter< mat< T, 1, 1 >, identity_t >

struct linalg::converter< mat< T, 2, 2 >, identity_t >

struct linalg::converter< mat< T, 3, 3 >, identity_t >

struct linalg::converter< mat< T, 4, 4 >, identity_t >

struct linalg::converter< std::array< T, 1 >, vec< T, 1 > >

struct linalg::converter< std::array< T, 2 >, vec< T, 2 > >

struct linalg::converter< std::array< T, 3 >, vec< T, 3 > >

struct linalg::converter< std::array< T, 4 >, vec< T, 4 > >

struct linalg::converter< vec< T, 1 >, std::array< T, 1 > >

struct linalg::converter< vec< T, 2 >, std::array< T, 2 > >

struct linalg::converter< vec< T, 3 >, std::array< T, 3 > >

struct linalg::converter< vec< T, 4 >, std::array< T, 4 > >

struct linalg::detail::any_compare

struct linalg::detail::any_compare< mat< T, M, 1 >, mat< T, M, 1 > >

struct linalg::detail::any_compare< mat< T, M, 2 >, mat< T, M, 2 > >

struct linalg::detail::any_compare< mat< T, M, 3 >, mat< T, M, 3 > >

struct linalg::detail::any_compare< mat< T, M, 4 >, mat< T, M, 4 > >

struct linalg::detail::any_compare< vec< T, 1 >, vec< T, 1 > >

struct linalg::detail::any_compare< vec< T, 2 >, vec< T, 2 > >

struct linalg::detail::any_compare< vec< T, 3 >, vec< T, 3 > >

struct linalg::detail::any_compare< vec< T, 4 >, vec< T, 4 > >

struct linalg::detail::apply

struct linalg::detail::apply< F, scalars_t< >, mat< A, M, N > >

struct linalg::detail::apply< F, scalars_t< >, mat< A, M, N >, mat< B, M, N > >

struct linalg::detail::apply< F, scalars_t< >, vec< A, M > >

struct linalg::detail::apply< F, scalars_t< >, vec< A, M >, vec< B, M > >

struct linalg::detail::apply< F, scalars_t< >, vec< A, M >, vec< B, M >, vec< C, M > >

struct linalg::detail::apply< F, scalars_t< A >, A, mat< B, M, N > >

struct linalg::detail::apply< F, scalars_t< A >, A, vec< B, M > >

struct linalg::detail::apply< F, scalars_t< A >, A, vec< B, M >, vec< C, M > >

struct linalg::detail::apply< F, scalars_t< A, B >, A, B, vec< C, M > >

struct linalg::detail::apply< F, scalars_t< A, C >, A, vec< B, M >, C >

struct linalg::detail::apply< F, scalars_t< A... >, A... >

struct linalg::detail::apply< F, scalars_t< B >, mat< A, M, N >, B >

struct linalg::detail::apply< F, scalars_t< B >, vec< A, M >, B >

struct linalg::detail::apply< F, scalars_t< B >, vec< A, M >, B, vec< C, M > >

struct linalg::detail::apply< F, scalars_t< B, C >, vec< A, M >, B, C >

struct linalg::detail::apply< F, scalars_t< C >, vec< A, M >, vec< B, M >, C >

struct linalg::detail::clamp

struct linalg::detail::empty

struct linalg::detail::getter

struct linalg::detail::getter< 0 >

struct linalg::detail::getter< 1 >

struct linalg::detail::getter< 2 >

struct linalg::detail::getter< 3 >

struct linalg::detail::lerp

struct linalg::detail::make_seq_impl

struct linalg::detail::make_seq_impl< A, 0 >

struct linalg::detail::make_seq_impl< A, 1 >

struct linalg::detail::make_seq_impl< A, 2 >

struct linalg::detail::make_seq_impl< A, 3 >

struct linalg::detail::make_seq_impl< A, 4 >

struct linalg::detail::max

struct linalg::detail::min

struct linalg::detail::op_add

struct linalg::detail::op_and

struct linalg::detail::op_cmp

struct linalg::detail::op_div

struct linalg::detail::op_eq

struct linalg::detail::op_ge

struct linalg::detail::op_gt

struct linalg::detail::op_int

struct linalg::detail::op_le

struct linalg::detail::op_lsh

struct linalg::detail::op_lt

struct linalg::detail::op_mod

struct linalg::detail::op_mul

struct linalg::detail::op_ne

struct linalg::detail::op_neg

struct linalg::detail::op_not

struct linalg::detail::op_or

struct linalg::detail::op_pos

struct linalg::detail::op_rsh

struct linalg::detail::op_sub

struct linalg::detail::op_un

struct linalg::detail::op_xor

struct linalg::detail::ord

struct linalg::detail::scalar_type

struct linalg::detail::scalar_type< mat< T, M, N > >

struct linalg::detail::scalar_type< vec< T, M > >

struct linalg::detail::scalars

struct linalg::detail::scalars<>

struct linalg::detail::select

struct linalg::detail::seq

struct linalg::detail::std_abs

struct linalg::detail::std_acos

struct linalg::detail::std_asin

struct linalg::detail::std_atan

struct linalg::detail::std_atan2

struct linalg::detail::std_ceil

struct linalg::detail::std_copysign

struct linalg::detail::std_cos

struct linalg::detail::std_cosh

struct linalg::detail::std_exp

struct linalg::detail::std_floor

struct linalg::detail::std_fmod

struct linalg::detail::std_log

struct linalg::detail::std_log10

struct linalg::detail::std_pow

struct linalg::detail::std_round

struct linalg::detail::std_sin

struct linalg::detail::std_sinh

struct linalg::detail::std_sqrt

struct linalg::detail::std_tan

struct linalg::detail::std_tanh

struct linalg::identity_t

struct linalg::mat

struct linalg::mat< T, M, 1 >

struct linalg::mat< T, M, 2 >

struct linalg::mat< T, M, 3 >

struct linalg::mat< T, M, 4 >

struct linalg::vec

struct linalg::vec< T, 1 >

struct linalg::vec< T, 2 >

struct linalg::vec< T, 3 >

struct linalg::vec< T, 4 >

struct std::hash< linalg::mat< T, M, 1 > >

struct std::hash< linalg::mat< T, M, 2 > >

struct std::hash< linalg::mat< T, M, 3 > >

struct std::hash< linalg::mat< T, M, 4 > >

struct std::hash< linalg::vec< T, 1 > >

struct std::hash< linalg::vec< T, 2 > >

struct std::hash< linalg::vec< T, 3 > >

struct std::hash< linalg::vec< T, 4 > >

class std::conditional::type

struct linalg::detail::scalars< T, U... >