Skip to content

Namespace ILLIXR

Namespace List > ILLIXR

Classes

Type 	Name

struct 	accel

struct 	camera_frame

class 	data_use_indicator
A helper class that lets one dynamically determine if some data gets used.

class 	dynamic_lib

class 	gen_guid
This class generates unique IDs.

struct 	hmd_physical_info

struct 	hologram_input

struct 	hologram_output

struct 	imu_cam_type

class 	phonebook
A service locator for ILLIXR.

class 	plugin
A dynamically-loadable plugin for Spindle.

struct 	pose_type

class 	reader_latest <typename event>
A handle which can read the latest event on a topic.

class 	record
This class represents a tuple of fields which get logged by record_logger .

class 	record_coalescer
Coalesces logs of the same type to be written back as a single-transaction.

class 	record_header
Schema of each record.

class 	record_logger
The ILLIXR logging service for structured records.

struct 	rendered_frame

struct 	rendered_frame_alt

struct 	rgb_depth_type

class 	runtime

class 	start_end_logger
a utility for logging to start and end times to a file

class 	switchboard
A manager for typesafe, threadsafe, named event-streams (called topics).

class 	threadloop
A reusable threadloop for plugins.

class 	writer <typename event>
A handle which can publish events to a topic.

class 	xlib_gl_extended_window

Public Types

Type 	Name

typedef plugin *(* 	plugin_factory

typedef std::chrono::time_point< std::chrono::system_clock > 	time_type

typedef unsigned long long 	ullong

typedef std::unique_ptr< void, std::function< void(void *)> > 	void_ptr

Public Attributes

Type 	Name

const record_header 	__plugin_start_header = /* multi line expression */

const record_header 	__threadloop_iteration_header = /* multi line expression */

Public Static Attributes

Type 	Name

std::chrono::milliseconds 	LOG_BUFFER_DELAY = {1000}

Public Functions

Type 	Name

runtime * 	runtime_factory (GLXContext appGLCtx)

Public Types Documentation

typedef plugin_factory

typedef plugin *(* ILLIXR::plugin_factory) (phonebook *);

typedef time_type

typedef std::chrono::time_point<std::chrono::system_clock> ILLIXR::time_type;

typedef ullong

typedef unsigned long long ILLIXR::ullong;

typedef void_ptr

using ILLIXR::void_ptr = typedef std::unique_ptr<void, std::function<void(void*)> >;

Public Attributes Documentation

variable __plugin_start_header

const record_header ILLIXR::__plugin_start_header;

variable __threadloop_iteration_header

const record_header ILLIXR::__threadloop_iteration_header;

Public Static Attributes Documentation

variable LOG_BUFFER_DELAY

std::chrono::milliseconds ILLIXR::LOG_BUFFER_DELAY;

Public Functions Documentation

function runtime_factory

runtime * ILLIXR::runtime_factory (
    GLXContext appGLCtx
) 

The documentation for this class was generated from the following file common/data_format.hpp